Re Create Map on locomotion change
