+++ app/javascript/packs/application.js
@@ -7,12 +7,9 @@ require("channels")
 // External imports
 import "bootstrap";
 
-// Sortable
-
-import Sortable from 'sortablejs';
-
 // Internal imports
-import { initMapboxes, initJourneyChange } from '../components/init_journey_change.js'
+import { initLocomotionChange } from '../components/init_locomotion_change.js'
+import { initMapboxes } from '../components/init_mapbox.js'
 import { initMinuteCreation } from '../components/init_minute_creation.js';
 import { initSpeechApi } from "../components/init_speech_api.js";
 import { initSortable } from "../plugins/init_sortable";
@@ -22,7 +19,7 @@ import { toggleMap } from "../components/toggle_map"
 import { modal } from "../components/modal"
 
 document.addEventListener('turbolinks:load', () => {
-  initJourneyChange();
+  initLocomotionChange();
   initMapboxes();
   initSpeechApi();
   initMinuteCreation();
